requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
scalar-fastapi==1.0.3
openai==1.54.3
anthropic==0.39.0
groq==0.13.0
httpx==0.27.2
python-dotenv==1.0.1
supabase==2.10.0
psycopg2-binary==2.9.10